improve mobile metrics
