+++ static/src/stylesheets/module/release-messages/_sign-in.scss
@@ -9,29 +9,41 @@ $color-sign-in-eb-callout: $news-garnett-highlight;
     display: flex;
     align-content: center;
     align-items: center;
-}
-
-.site-message__message.site-message__message--sign-in {
-    max-width: 100%;
+    overflow-y: auto;
+    .gs-container {
+        margin: auto;
+    }
+    .site-message__inner {
+        background: $color-sign-in-eb-main;
+        color: $garnett-neutral-1;
+        animation: pop-in .5s cubic-bezier(.175, .885, .32, 1.275);
+        max-width: 310px;
+        padding: 0;
+        position: relative;
+        margin: $gs-baseline auto;
+    }
+    .site-message__message {
+        max-width: 100%;
+    }
+    .site-message__copy {
+        padding: 0;
+    }
 }
 
 .site-message--sign-in-container {
-    background: $color-sign-in-eb-main;
-    color: $garnett-neutral-1;
-    animation: pop-in .5s cubic-bezier(.175, .885, .32, 1.275);
-    max-width: 310px;
-    padding: 0;
-    position: relative;
     .site-message--sign-in__header,
     .site-message--sign-in__body,
     .site-message--sign-in__buttons {
-        padding: $gs-gutter;
+        padding: $gs-baseline * 1.5;
     }
 }
 
 .site-message--sign-in__header {
     @include fs-headline(3);
-    @media all and(min-height: 40em) {
+    @media all and(min-height: 28em) {
+        @include fs-headline(4, true);
+    }
+    @media all and(min-height: 33em) {
         @include fs-headline(5, true);
     }
     .site-message--sign-in__header-msg {
@@ -91,7 +103,6 @@ a.site-message--sign-in-cta {
     justify-content: space-between;
     transition: .125s;
     box-sizing: border-box;
-    text-decoration: none;
 }
 
 .site-message--sign-in__buttons {
@@ -107,11 +118,10 @@ a.site-message--sign-in__why {
     @include fs-textSans(3);
     display: block;
     background: $color-sign-in-eb-alt;
-    color: #fff;
+    color: #ffffff;
     text-decoration: none;
-    padding: $gs-gutter/4 $gs-gutter;
+    padding: $gs-baseline/3 $gs-baseline * 1.5;
 }
-
 
 .site-message--sign-in__dismiss {
     @include fs-textSans(3);
@@ -119,8 +129,22 @@ a.site-message--sign-in__why {
     background: $color-sign-in-eb-callout;
     color: $garnett-neutral-1;
     padding: $gs-gutter/4;
-    border: none;
+    border: 0;
     position: absolute;
     top: $gs-gutter * -.5;
     right: $gs-gutter * -.5;
+}
+
+.site-message--sign-in-container .site-message--sign-in__body {
+    @include fs-bodyCopy(1);
+    list-style: none;
+    margin: 0;
+    padding-top: 0;
+    padding-bottom: 0;
+    li {
+        margin: 0 0 $gs-baseline;
+        strong {
+            display: block;
+        }
+    }
 }
